refactor(storage): use async fn in implementations
Simplify MemoryStorage, CorruptingStorage, and FaultyStorage
implementations of the Storage trait by utilizing the clean native
async fn syntax instead of manually returning impl Future + Send
blocks. Remove obsolete clippy allow attributes and redundant casts.

+++ fuzz/fuzz_targets/proof_mutation.rs
@@ -66,7 +66,7 @@ fuzz_target!(|input: Input| {
         let mut log = Log::new(MemoryStorage::new());
         log.add_algorithm(0, Box::new(FuzzHasher)).await.unwrap();
         for i in 0..n {
-            log.append(&(i as u64).to_le_bytes()).await.unwrap();
+            log.append(&i.to_le_bytes()).await.unwrap();
         }
 
         let tree_size = log.tree_size(0).await.unwrap();
@@ -108,7 +108,7 @@ fuzz_target!(|input: Input| {
             let mut old_log = Log::new(MemoryStorage::new());
             old_log.add_algorithm(0, Box::new(FuzzHasher)).await.unwrap();
             for i in 0..mid {
-                old_log.append(&(i as u64).to_le_bytes()).await.unwrap();
+                old_log.append(&i.to_le_bytes()).await.unwrap();
             }
             let old_root = old_log.root(0).unwrap();
